+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   - The Duly Authorized Representative can now review the entire agreement before signing
   - Agreement changes are now audited so that reviewers can view changes more easily between resubmissions
   - Principal reviewers can now modify the requested datasets during the review process
+  - Principal reviewers can now export information on all agreements as a CSV
 - **Challenge Changes**
   - Added an export for the Flow Limitation challenge
 - **Tool Changes**